tokens: Add repo subset and clarify permissions
We now also have a list of repos in the token, where '' means any.
Additionally we make the prefix part of the token non-optional.

However both prefix and repo are optional when we're subsetting, an
if not set are directly inherited from the current claim.

This later is needed because we may not know what the prefix
of our current token is on the client side when subsetting, but we still
want to be able to subset other things.