Split out repo update from publish job
Also, we reuse existing non-started repo update job for the
same repo. This way multiple publish ops can share a
single update. New update jobs depend on any previous ones
to avoid them being parallelized.

We point to the status page for the update job in the logs so
that users can follow the logs.