jobs: Better handling of delta deletion
Instead of just deleting deltas we move them from deltas/ to tmp/deltas/
in the repo, and then expose such files as if they were in deltas/
when accessing the repo. This way old operations can still (atomically)
see the old deltas, but they are not part of the new summary.