deltas: Clean up worker handling
Instead of having a shared vector and a generic priority field we
split this up into the fallback local_worker and a remote_workers
list.

This also allows us to do some generic Addr<A> stuff and drop
the weird WorkerWrapper thing.

Also, we new randomize picking of remote workers to spread the load.