jobs: Add DeployCollectionID key to flatpakref
In preparation for Flathub deploying collection IDs to users, add the
DeployCollectionID key to generated flatpakref files. Although only
Flatpak >= 1.0.6 will honor it, older versions will ignore it without
producing an error.

This is conditional on a new config key so that it can be set for the
main repo but not build repos.

The addition of DeployCollectionID to the flatpakrepo file and
ostree.deploy-collection-id to the repo metadata will have to be made
manually.