repo: Allow authentication for private repos
Add a private boolean to the repo configuration to require
authentication when accessing the repository. To support this, the
TokenParser middleware is added to for repo requests. If the repo has
been declared private, then the client must authenticate with a valid
JWT token. Currently the authorization is all or nothing, but later this
will be extended to match claims to requested refs.