ostree: Get bindings from commit objects and delta superblocks
In order to match the requested commit or delta superblock to the
allowed refs in the token, we need to get the ref and collection
bindings out of the commit object. This adds some helpers to return a
CommitBindings struct containing these values.

While I'm sure it's possible to do all of this in native rust, I really
didn't want to rewrite the GVariant parser here for the complex types
used commit and superblock variants. This pulls in glib and glib_sys in
order to call out to the various g_variant* APIs directly.