repo: Verify tokens for commits with configured token-types
When getting a commit or superblock we load it and check for
xa.token-type in the metadata (falling back to the configured default).
If the token type is one of the configured ones that require authentication
we extract the xa.ref from the commit object and fail if the id part
of that ref doesn't match a prefix in the supplied token.