tokens: Make some jwt token parts optional
We allow missing prefixes, repos and name, as these are not going to
be needed for repo specific tokens.

The name is just a debug thing and this is not sensitive. The other
ones don't really affect anything. They just let you avoid specifying
no key instead of an empty array.