tokens: Use simpler tokens for repo access
The API token checks are as before, but the repo checks now just use
"scope" to mean a list of id prefixes, and "sub" the user id (although
this is not used atm).

There is no need for all the other complexity in this case, and means
we can create more standardized tokens in the authenticator.