Allow builds to be private
Allow build repos to be set as public or private. If no app ID is set,
the default is public as before, but if an app ID *is* set, the default
is private (but in either case can be overridden). flat-manager-client gains
`--public_download`/`--no_public_download` flags to control this
behavior.

A private build requires a build token in order to download from it. A new scope
is added, "download", which allows listing and downloading builds.